Move secret generation
